+++ src/sqlitely/conf.py
@@ -29,7 +29,7 @@ except ImportError: wx = None
 
 """Program title, version number and version date."""
 Title = "SQLitely"
-Version = "2.1.dev113"
+Version = "2.1.dev114"
 VersionDate = "14.08.2023"
 
 if getattr(sys, "frozen", False):